dbClient/requirements.txt
cffi==1.3.1
mysql-connector-python==2.0.4
pycparser==2.14
tabulate==0.7.5
wsgiref==0.1.2
